+++ syemessenger-api/src/main/java/io/syemessenger/api/room/GetRoomMembersResponse.java
@@ -9,7 +9,7 @@ public class GetRoomMembersResponse {
   private List<AccountInfo> accountInfos;
   private Integer offset;
   private Integer limit;
-  private Integer totalCount;
+  private Long totalCount;
 
   public List<AccountInfo> accountInfos() {
     return accountInfos;
@@ -38,11 +38,11 @@ public class GetRoomMembersResponse {
     return this;
   }
 
-  public Integer totalCount() {
+  public Long totalCount() {
     return totalCount;
   }
 
-  public GetRoomMembersResponse totalCount(Integer totalCount) {
+  public GetRoomMembersResponse totalCount(Long totalCount) {
     this.totalCount = totalCount;
     return this;
   }
